Adds init command to scaffold SiteKit project
Adds an "init" command to the SiteKit CLI to initialize a new SiteKit project with sample YAML files.

This command creates the necessary directory structure (.sitekit/[site_name]) and downloads pre-defined YAML files for components, composition, dictionary, page types, and site settings.

The sitesettings.yaml file is then updated with the tenant and site names provided as input to the command.